requirements.txt
python-telegram-bot>=20.0
FastAPI>=0.104.0
uvicorn>=0.24.0
Flask==3.0.2
python-dotenv>=1.0.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.2
aiohttp>=3.8.0
requests==2.31.0
firebase-admin>=6.0.0
google-cloud-storage>=2.10.0
google-cloud-texttospeech>=2.14.0
stripe>=6.0.0
python-ton>=0.3.0
google-cloud-firestore>=2.11.0